snap: Include default sdplayer.ini
This adds the sdplayer.ini from the previous commit to what the snap is
bundled with, resolving #2. This also switches away from using a custom
build step to using `organize` to ship the executable -- organize wasn't
working before because apparently it's based on the install directory,
not the build directory
(https://snapcraft.io/docs/parts-lifecycle#heading--parts-directories).